update version of project
