+++ src/components/common/ProductBox/ProductBox.js
@@ -1,41 +1,44 @@
-import React, { useState } from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import clsx from 'clsx';
+import { useDispatch, useSelector } from 'react-redux';
+import { addToFavorites, removeFromFavorites } from '../../../redux/productsRedux';
 import styles from './ProductBox.module.scss';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faExchangeAlt, faShoppingBasket } from '@fortawesome/free-solid-svg-icons';
-import { faHeart } from '@fortawesome/free-regular-svg-icons';
+import {
+  faStar,
+  faExchangeAlt,
+  faShoppingBasket,
+} from '@fortawesome/free-solid-svg-icons';
+import { faStar as farStar, faHeart } from '@fortawesome/free-regular-svg-icons';
 import Button from '../Button/Button';
-import StarsRating from '../../features/StarsRating/StarsRating';
-import { addProductToCompare } from '../../../redux/compareRedux';
-import { useDispatch } from 'react-redux';
+import initialState from '../../../redux/initialState';
 
-const ProductBox = ({ name, price, promo, stars, isFavorite, isCompare, id, ownRating, oldPrice  }) => {
-  const buttonFavoriteActive = clsx('outline', {
-    [styles.favorite]: isFavorite,
-  });
-  const buttonCompareActive = clsx('outline', {
-    [styles.favorite]: isCompare,
-  });
-  const [isHovered, setIsHovered] = useState(false);
+const ProductBox = ({ id, name, price, promo, stars }) => {
   const dispatch = useDispatch();
-  const product = { name };
+  const favorites = useSelector(state => state.products.favorites || {});
+  console.log('Favorites:', favorites);
 
-  const addToCompare = () => {
-    dispatch(addProductToCompare(product));
+  useEffect(() => {
+    if (Object.keys(favorites).length === 0) {
+      dispatch(addToFavorites({ id: initialState.products[0].id }));
+    }
+  }, [dispatch, favorites, id]);
+
+  const isFavorite = favorites[id];
+
+  const handleToggleFavorite = () => {
+    if (isFavorite) {
+      dispatch(removeFromFavorites({ id }));
+    } else {
+      dispatch(addToFavorites({ id }));
+    }
   };
   return (
-    <div
-      className={styles.root}
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
-    >
+    <div className={styles.root}>
       <div className={styles.photo}>
+        <img src={`images/beds/${name}.jpg`} alt={name} />
         {promo && <div className={styles.sale}>{promo}</div>}
-        <div
-          className={styles.buttons}
-          style={isHovered === true ? { opacity: 1 } : { opacity: 0 }}
-        >
+        <div className={styles.buttons}>
           <Button variant='small'>Quick View</Button>
           <Button variant='small'>
             <FontAwesomeIcon icon={faShoppingBasket}></FontAwesomeIcon> ADD TO CART
@@ -45,7 +48,6 @@ const ProductBox = ({ name, price, promo, stars, isFavorite, isCompare, id, ownR
       <div className={styles.content}>
         <h5>{name}</h5>
         <div className={styles.stars}>
-
           {[1, 2, 3, 4, 5].map(i => (
             <a key={i} href='#'>
               {i <= stars ? (
@@ -55,40 +57,39 @@ const ProductBox = ({ name, price, promo, stars, isFavorite, isCompare, id, ownR
               )}
             </a>
           ))}
-
-          <StarsRating stars={stars} id={id} ownRating={ownRating} />
         </div>
       </div>
       <div className={styles.line}></div>
       <div className={styles.actions}>
         <div className={styles.outlines}>
-          <Button variant='outline' className={buttonFavoriteActive}>
+          <Button
+            variant='outline'
+            onClick={handleToggleFavorite}
+            className={isFavorite ? styles.favorite : ''}
+          >
             <FontAwesomeIcon icon={faHeart}>Favorite</FontAwesomeIcon>
           </Button>
-          <Button variant='outline' onClick={addToCompare}  className={buttonCompareActive}>
+          <Button variant='outline'>
             <FontAwesomeIcon icon={faExchangeAlt}>Add to compare</FontAwesomeIcon>
           </Button>
         </div>
         <div className={styles.price}>
-        {oldPrice ? <span className={styles.oldPrice}>${oldPrice}</span> : ''}
-        <Button noHover variant='small'>
-          $ {price}
-        </Button>
+          <Button noHover variant='small'>
+            $ {price}
+          </Button>
+        </div>
       </div>
     </div>
   );
 };
+
 ProductBox.propTypes = {
   children: PropTypes.node,
+  id: PropTypes.string,
   name: PropTypes.string,
   price: PropTypes.number,
   promo: PropTypes.string,
   stars: PropTypes.number,
-  oldPrice: PropTypes.number,
-  isFavorite: PropTypes.bool,
-  isCompare: PropTypes.bool,
-  id: PropTypes.string,
-  ownRating: PropTypes.number,
 };
 
 export default ProductBox;
