+++ src/redux/productsRedux.js
@@ -1,4 +1,6 @@
 const createActionName = actionName => `app/products/${actionName}`;
+const ADD_TO_FAVORITES = createActionName('ADD_TO_FAVORITES');
+const REMOVE_FROM_FAVORITES = createActionName('REMOVE_TO_FAVORITES');
 const UPDATE_YOUR_STARS_RATE = createActionName('UPDATE_YOUR_STARS_RATE');
 
 /* selectors */
@@ -10,14 +12,40 @@ export const getNew = ({ products }) =>
 
 export const allPromotional = ({ promotional }) => promotional;
 export const addStarsRating = payload => ({ type: UPDATE_YOUR_STARS_RATE, payload });
+export const addToFavorites = payload => ({ type: ADD_TO_FAVORITES, payload });
+export const removeFromFavorites = payload => ({
+  type: REMOVE_FROM_FAVORITES,
+  payload,
+});
 
 /* reducer */
-export default function reducer(statePart = [], action = {}) {
+const initialState = {
+  favorites: {},
+};
+
+export default function reducer(statePart = initialState, action = {}) {
   switch (action.type) {
-    case UPDATE_YOUR_STARS_RATE:
-      return statePart.map(product =>
-        product.id === action.payload.id ? { ...product, ...action.payload } : product
-      );
+    case ADD_TO_FAVORITES: {
+      const { id } = action.payload;
+      const updatedFavorites = { ...statePart.favorites, [id]: true };
+
+      return {
+        ...statePart,
+        favorites: updatedFavorites,
+      };
+    }
+
+    case REMOVE_FROM_FAVORITES: {
+      const { id } = action.payload;
+      const updatedFavorites = { ...statePart.favorites };
+      delete updatedFavorites[id];
+
+      return {
+        ...statePart,
+        favorites: updatedFavorites,
+      };
+    }
+
     default:
       return statePart;
   }
